requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlmodel>=0.0.8
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=5.0.1
psycopg2-binary>=2.9.9
alembic>=1.12.0
httpx>=0.25.0
